Check

+++ src/main/java/Logger/LoggerSvc.java
@@ -1,19 +1,23 @@
 package Logger;
 
+import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
+import java.nio.file.StandardOpenOption;
 import java.sql.SQLOutput;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
 
 public class LoggerSvc {
 
-    OutputStream object;
-
+    FileOutputStream object;
+    long charCount=0;
     LoggerSvc(String fileName) {
         try {
-            object = new FileOutputStream("./files/abc.txt");
+            File file = new File(fileName);
+            file.createNewFile();
+            object = new FileOutputStream(fileName);
 
 
         } catch (Exception e) {
@@ -23,13 +27,23 @@ public class LoggerSvc {
 
 
     public void addLog(String data, String processId) throws ExecutionException, InterruptedException, IOException {
+        this.charCount += (processId+data).length();
         object.write((processId+data).getBytes());
-        CompletableFuture.runAsync(() -> flushAll()).get();
+        System.out.println("Add log start");
+
+        if(charCount>1000){
+            CompletableFuture.runAsync(() -> flushAll()).get();
+        }
+        System.out.println("Add log end");
     }
 
     public void flushAll() {
         try {
+            System.out.println("Flush called");
             object.flush();
+            this.charCount =0;
+//            object.close();
+            System.out.println("");
         } catch (IOException e) {
             e.printStackTrace();
         }
